Counted most popular morphemes

+++ table.h
@@ -20,23 +20,23 @@ public:
 		prefexmp(20), suffexmp(10), morfexmp(20), wordexmp(maxdist), distexmp(40)
 	{
 		tab=new int**[maxsize];
-		for (int i=0; i<maxsize; ++i){
+		for (int i = 0; i < maxsize; ++i){
 			tab[i] = new int*[i];
-			for (int j=0; j<i; ++j){
+			for (int j = 0; j < i; ++j){
 				tab[i][j] = new int[2];
 			}
 		}
 		size = 0;
-		for (int i=0; i<10; ++i){
+		for (int i = 0; i < 10; ++i){
 			prefstat[i] = 0;
 		}
-		for (int i=0; i<10; ++i)
+		for (int i = 0; i < 10; ++i)
 			suffstat[i] = 0;
-		for (int i=0; i<20; ++i)
+		for (int i = 0; i < 20; ++i)
 			morfstat[i] = 0;
-		for (int i=0; i<40; ++i)
+		for (int i = 0; i < 40; ++i)
 			diststat[i] = 0;		
-		for (int i=0; i<maxdist; ++i)
+		for (int i = 0; i < maxdist; ++i)
 			wordstat[i] = 0;
 		prefstat.reset();
 		suffstat.reset();
@@ -46,18 +46,18 @@ public:
 		cnt = 0;
 	}
 	~table(){
-		for (int i=0; i<maxsize; ++i){
-			for (int j=0; j<i; ++j){
-				delete [] tab[i][j];
+		for (int i = 0; i < maxsize; ++i){
+			for (int j = 0; j < i; ++j){
+				delete[] tab[i][j];
 			}
 			delete [] tab[i];
 		}
 		delete [] tab;
 	}
 	int operator()(int i, int j, int k) const{
-		if (i>j)
+		if (i > j)
 			return tab[i][j][k];
-		else if (i<j)
+		else if (i < j)
 			return tab[j][i][k];
 		else
 			return 0;
@@ -101,34 +101,34 @@ public:
 					tab[i][j][1] = dist(s1.suff, s2.suff);
 					int dist12 = dist(s1.word, s2.word);
 					++wordstat[dist12];
-					if (wordexmp[dist12].size() < maxexmp){
+					if ((int)wordexmp[dist12].size() < maxexmp){
 						wordexmp[dist12].push_back(s1.word);
 						wordexmp[dist12].push_back(s2.word);
 					}
 					int distortion = (int)((double) dist12 / (s1.word.size * s2.word.size) * 100);
 					diststat[distortion]++;
-					if (distexmp[distortion].size() < maxexmp){
+					if ((int)distexmp[distortion].size() < maxexmp){
 						distexmp[distortion].push_back(s1.word);
 						distexmp[distortion].push_back(s2.word);
 						//cerr << distortion << ' ' << (double) dist12 / (s1.word.size * s2.word.size) << endl;
 					}
 				}else{
-					tab[i][j][0]=9;
-					tab[i][j][1]=9;
+					tab[i][j][0] = 9;
+					tab[i][j][1] = 9;
 					++wordstat[wordstat.maxlen - 1];
 				}				
 				++prefstat[tab[i][j][0]];
-				if (prefexmp[tab[i][j][0]].size() < maxexmp){
+				if ((int)prefexmp[tab[i][j][0]].size() < maxexmp){
 					prefexmp[tab[i][j][0]].push_back(s1.word);
 					prefexmp[tab[i][j][0]].push_back(s2.word);
 				}
 				++suffstat[tab[i][j][1]];
-				if (suffexmp[tab[i][j][1]].size() < maxexmp){
+				if ((int)suffexmp[tab[i][j][1]].size() < maxexmp){
 					suffexmp[tab[i][j][1]].push_back(s1.word);
 					suffexmp[tab[i][j][1]].push_back(s2.word);
 				}
 				++morfstat[tab[i][j][0]+tab[i][j][1]];
-				if (morfexmp[tab[i][j][0]+tab[i][j][1]].size() < maxexmp){
+				if ((int)morfexmp[tab[i][j][0]+tab[i][j][1]].size() < maxexmp){
 					morfexmp[tab[i][j][0]+tab[i][j][1]].push_back(s1.word);
 					morfexmp[tab[i][j][0]+tab[i][j][1]].push_back(s2.word);
 				}
